♻️ refactor:  WidgetとControllerを分離
Issue #285
