+++ lib/presentation/pages/repo/components/search_repos_sort.dart
@@ -5,29 +5,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../../../../domain/repositories/app_data/app_data_repository.dart';
-import '../../../../domain/repositories/app_data/entities/app_data_key.dart';
 import '../../../../domain/repositories/repo/entities/search_repos_sort.dart';
 import '../../../../utils/localizations/strings.g.dart';
 import '../../../../utils/logger.dart';
-import '../../../controllers/app_data.dart';
+import '../../../controllers/search_repos_sort.dart';
 import 'search_repos_order.dart';
-
-/// リポジトリ検索用ソート値プロバイダー
-final searchReposSortProvider =
-    StateNotifierProvider<SearchReposSortController, SearchReposSort>(
-  (ref) => SearchReposSortController(
-    appDataRepository: ref.watch(appDataRepositoryProvider),
-  ),
-  name: 'searchReposSortProvider',
-);
-
-/// リポジトリ検索用ソート値コントローラー
-class SearchReposSortController extends AppDataController<SearchReposSort> {
-  SearchReposSortController({
-    required super.appDataRepository,
-  }) : super(appDataKey: AppDataKey.searchReposSort);
-}
 
 /// リポジトリ検索用ソート選択ボトムシート
 class SearchReposSortSelectorBottomSheet extends ConsumerWidget {
