+++ lib/presentation/pages/repo/repo_search_page.dart
@@ -5,7 +5,7 @@
 import 'package:flutter/material.dart';
 
 import '../../widgets/search_app_bar.dart';
-import 'components/query_histories.dart';
+import 'components/query_history.dart';
 import 'components/search_repos_query.dart';
 import 'components/search_repos_sort.dart';
