+++ lib/utils/routing/router.dart
@@ -9,9 +9,9 @@ import 'package:go_router/go_router.dart';
 import 'package:responsive_framework/responsive_wrapper.dart';
 
 import '../../domain/repositories/repo/entities/repo.dart';
+import '../../presentation/controllers/selected_repo.dart';
 import '../../presentation/pages/error/error_page.dart';
 import '../../presentation/pages/repo/avatar_preview_page.dart';
-import '../../presentation/pages/repo/components/selected_repo.dart';
 import '../../presentation/pages/repo/repo_index_page.dart';
 import '../../presentation/pages/repo/repo_search_page.dart';
 import '../../presentation/pages/repo/repo_view_page.dart';
