+++ test/presentation/pages/repo/components/avatar_preview_view_test.dart
@@ -4,8 +4,8 @@
 
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
+import 'package:github_search/presentation/controllers/selected_repo.dart';
 import 'package:github_search/presentation/pages/repo/components/avatar_preview_view.dart';
-import 'package:github_search/presentation/pages/repo/components/selected_repo.dart';
 import 'package:photo_view/photo_view.dart';
 
 import '../../../../test_utils/golden_testing_tools.dart';
