+++ test/presentation/pages/repo/components/repo_detail_view_test.dart
@@ -7,9 +7,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:github_search/domain/repositories/repo/entities/repo.dart';
+import 'package:github_search/presentation/controllers/selected_repo.dart';
 import 'package:github_search/presentation/pages/repo/components/readme_markdown.dart';
 import 'package:github_search/presentation/pages/repo/components/repo_detail_view.dart';
-import 'package:github_search/presentation/pages/repo/components/selected_repo.dart';
 import 'package:github_search/presentation/widgets/cached_circle_avatar.dart';
 import 'package:github_search/presentation/widgets/error_view.dart';
 import 'package:github_search/utils/routing/router.dart';
