+++ test/presentation/pages/repo/components/selected_repo_test.dart
@@ -7,7 +7,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:github_search/domain/repositories/repo/entities/repo.dart';
 import 'package:github_search/infrastructure/github/repo/json_objects/repo.dart';
 import 'package:github_search/infrastructure/github/repo/repo_repository.dart';
-import 'package:github_search/presentation/pages/repo/components/selected_repo.dart';
+import 'package:github_search/presentation/controllers/selected_repo.dart';
 
 import '../../../../test_utils/test_agent.dart';
 import '../../../../test_utils/utils.dart';
