+++ test/presentation/pages/repo/repo_search_page_test.dart
@@ -6,7 +6,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:github_search/domain/repositories/query_history/entities/query_history.dart';
 import 'package:github_search/domain/repositories/query_history/query_history_repository.dart';
-import 'package:github_search/presentation/pages/repo/components/query_histories.dart';
+import 'package:github_search/presentation/pages/repo/components/query_history.dart';
 import 'package:github_search/presentation/pages/repo/components/search_repos_query.dart';
 import 'package:github_search/presentation/pages/repo/components/search_repos_sort.dart';
 import 'package:github_search/presentation/pages/repo/repo_search_page.dart';
